libraries/sensor/hi3516ev200: fuse sony_imx335 and sony_imx335_fpv into one driver
The two IMX335 drivers shared the same sensor object (stSnsImx335Obj) so
they couldn't coexist in firmware — users had to pick one via the INI's
DllFile line. The _fpv variant was a strict superset of stock plus extra
high-fps modes (binning, 1080p cropped, 1520p, boost-1944p), so it's
adopted as canonical.

Mode 8 = IMX335_4M_25FPS_10BIT_WDR (2560×1440 WDR @ 25fps) was the only
mode unique to the stock driver. It's restored here at a fresh slot
(rather than reusing Mode 2, which _fpv repurposed for 1520P_10BIT) so
existing _fpv configurations with Isp_SnsMode=2 keep their meaning.
Mode 7 stays reserved for the in-flight flexible-crop work (#98).

Verified on openipc-hi3516ev300 + IMX335:

  Mode 0  2592x1944 linear  @30  30/30 fps  PASS
  Mode 1  2592x1944 WDR     @30  22/30 fps  (sensor ceiling, matches stock fpv)
  Mode 2  2592x1520 linear  @30  28/30 fps  PASS
  Mode 3  2592x1520 WDR     @30  30/30 fps  PASS
  Mode 4  1296x972 binning  @60  59/60 fps  PASS
  Mode 5  1920x1080 cropped @60  59/60 fps  PASS
  Mode 6  1944p boost       @40  35/40 fps  (sensor ceiling, matches stock fpv)
  Mode 8  2560x1440 WDR     @25  25/25 fps  PASS (restored from stock)

Companion firmware change required: presets referencing libsns_imx335_fpv.so
or imx335_fpv.ini need to switch to libsns_imx335.so / imx335.ini. A
compat symlink in the buildroot install rule keeps existing user configs
working through an upgrade.
