+++ libraries/sensor/hi3516ev200/sony_imx335/imx335.ini
@@ -1,7 +1,7 @@
 [sensor]
 Sensor_type=stSnsImx335Obj
 Mode=WDR_MODE_NONE
-DllFile=libsns_imx335_fpv.so
+DllFile=libsns_imx335.so
 
 [mode]
 input_mode=INPUT_MODE_MIPI
